WIP fix all mypy errors; check data handling and upload next
